Install atom and rust more easly

+++ install-scripts/install-atom-plugin.sh
@@ -1,14 +1,4 @@
 #!/usr/bin/env bash
-
-# Set nightly channel in atom config file for 'ide-rust' plugin.
-#
-# $1
-set_channel_in_atom_editor() {
-  filename="${HOME}/.atom/config.cson"
-  local line_number=$(cat "${filename}" | grep -n 'rlsToolchain' | cut -d ':' -f 1)
-  sed -i ${line_number}'s/.*/    rlsToolchain: "'$1'"/' "${filename}"
-}
-
 REALPATH="$(realpath $0)"
 BASEDIR="$(dirname ${REALPATH})"
 
@@ -24,8 +14,6 @@ if [ -z "$(command -v apm)" ]; then
   exit 1
 fi
 
-CHANNEL=$(cat "${TMP_RUST_CHANNEL}")
-
 for pck_name in ${ATOM_PACKAGE}; do
   # Check if package is installed
   IS_INSTALLED=$(apm list --installed --bare | grep "${pck_name}")
@@ -37,9 +25,10 @@ for pck_name in ${ATOM_PACKAGE}; do
   fi
 done
 
-cp ${BASEDIR}/template/config.cson ~/.atom/config.cson
-cp ${BASEDIR}/template/terminal-commands.json ~/.atom/terminal-commands.json
+if [ ! -f "~/.atom/config.cson" ]; then
+  cp ${BASEDIR}/template/config.cson ~/.atom/config.cson
+fi
 
-set_channel_in_atom_editor "${CHANNEL}"
-
-rm -rf "${TMP_RUST_CHANNEL}"
+if [ ! -f "~/.atom/terminal-commands.json" ]; then
+  cp ${BASEDIR}/template/terminal-commands.json ~/.atom/terminal-commands.json
+fi
